# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Rho Daemon API
  description: |
    HTTP API served by the rho daemon on port 4590.
    Used by sparrow, robin, and external integrations.
    The daemon must be running (`rho daemon start`) before calling these endpoints.

    This spec is kept in lockstep with the registered routes in
    internal/daemon; the daemon test suite fails if they drift.
  version: "0.1.0"
  license:
    name: MIT
    url: https://github.com/GrayCodeAI/rho/blob/main/LICENSE
  contact:
    url: https://github.com/GrayCodeAI/rho
servers:
  - url: http://127.0.0.1:4590
    description: Local daemon (default port)
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ReviewRequest:
      type: object
      required: [sha]
      properties:
        sha:
          type: string
          description: Hex git SHA (7-40 chars) to review
        background:
          type: boolean
        model:
          type: string
        concerns:
          type: string
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sha:
          type: string
        status:
          type: string
        message:
          type: string
    ReviewStatusResponse:
      type: object
      properties:
        status:
          type: string
          description: Output of `rho review status`
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
...
paths:
...
  /v1/review:
    post:
      operationId: createReview
      tags: [review]
      summary: Trigger an asynchronous code review of a commit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
      responses:
        "200":
          description: Review accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/review/status:
    get:
      operationId: getReviewStatus
      tags: [review]
      summary: Get current review status
      responses:
        "200":
          description: Review status text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewStatusResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Status command failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"